+++ src/com/braintreegateway/CreditCardVerification.java
@@ -13,17 +13,13 @@ public class CreditCardVerification {
     private String status;
 
     public CreditCardVerification(NodeWrapper node) {
-        node = node.findFirst("verification");
-
-        if (node != null) {
-            this.avsErrorResponseCode = node.findString("avs-error-response-code");
-            this.avsPostalCodeResponseCode = node.findString("avs-postal-code-response-code");
-            this.avsStreetAddressResponseCode = node.findString("avs-street-address-response-code");
-            this.cvvResponseCode = node.findString("cvv-response-code");
-            this.processorResponseCode = node.findString("processor-response-code");
-            this.processorResponseText = node.findString("processor-response-text");
-            this.status = node.findString("status");
-        }
+        this.avsErrorResponseCode = node.findString("avs-error-response-code");
+        this.avsPostalCodeResponseCode = node.findString("avs-postal-code-response-code");
+        this.avsStreetAddressResponseCode = node.findString("avs-street-address-response-code");
+        this.cvvResponseCode = node.findString("cvv-response-code");
+        this.processorResponseCode = node.findString("processor-response-code");
+        this.processorResponseText = node.findString("processor-response-text");
+        this.status = node.findString("status");
     }
 
     public String getAvsErrorResponseCode() {