+++ src/com/braintreegateway/Pager.java
@@ -1,5 +1,5 @@
 package com.braintreegateway;
 
 public interface Pager<T> {
-    public PagedCollection<T> getPage(int page);
+    public ResourceCollection<T> getPage(int page);
 }